Add files via upload
Respiration data from day5 of experiment. Treated results are on "Feuil3".
There are 2 results: one is from the fitting of the respiration data to a linear model and then, based on intercept and slope, calculation of the predicted O2 concentration after 24h. The second is just the difference actually measured with the optode.